requirements.txt
streamlit
langchain
langgraph
openai
yfinance
pandas
plotly
ta
requests
python-dotenv
langchain_groq